requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
ultralytics==8.0.196
python-telegram-bot==20.6
aiohttp==3.9.1
requests==2.31.0
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
numpy==1.24.3
pillow==10.1.0

